sys-kernel/coreos-sources: remove Xen-PCI-revert patch
There was a kernel regression on Xen HVM with regard to MSI interrupts that
affected certain AWS instances (m4 and similar). We reverted the patch that
broke networking, but in the meantime upstream found the actual cause and
provided a proper fix which is part of 5.10.114. Remove the obsolete patch.

Link: https://lore.kernel.org/all/20220504153024.088119748@linuxfoundation.org/