profiles: Remove unnecessary package.provided for app-arch/xz-utils
The circular dep used to be:

app-arch/xz-utils
libtool.eclass
app-portage/elt-patches
app-arch/xz-utils

The circular dep should be gone now, because app-arch/xz-utils is a
BDEPEND, so for building production images the package from SDK will
be used. For building SDK the package from seed SDK should be used.