profiles: Do not install repos.conf from sys-apps/portage
The package installs a default repos.conf in
/usr/share/portage/config/repos.conf. Normally it contained a default
configuration for gentoo repository, but we have a modification in the
ebuild that clobbers the contents of the file with a comment
line. Which means that this configuration file is empty. We do so,
because we are putting our own configuration in /etc anyway.

To minimize the amount of modifications put into the ebuild, we can
add /usr/share/portage/config/repos.conf into INSTALL_MASK. This is
one modification less and it's putting us closer to moving
sys-apps/portage to portage-stable.