profiles: Disable binpkg-multi-instance feature
The binpkg-multi-instance feature seems to be about having multiple
binary packages of a single source package, in case those packages
were built with different USE flags and stuff.

So far we have not used this feature, and it looks like that the
update of portage has made this feature to be enabled by default. The
enabling broke the image build, because this feature makes portage to
use a bit different directory structure to store binary packages. So
instead of storing sys-devel/gcc binary package in
${PKGDIR}/sys-devel/gcc-${VERSION}.tbz2, the binary package is stored
in ${PKGDIR}/sys-devel/gcc/gcc-${VERSION}-${BUILD_ID}.xpak.

Maybe we could consider using this feature at some point in future,
but for now disable it to keep building things working.