profiles: Disable split-usr for generic images completely
We do specify -split-usr in make.defaults for the USE variable but
it's not enough - the base profile in portage-stable forces this flag
and this overrides our defaults. As a workaround, we were using
package.use.force to force-disable split-usr for selected
packages. Now, with addition of more split-usr-using packages in
app-alternatives the list in package.use.force would grow. Instead of
listing all the packages having split-usr in their IUSE, use the
use.force and use.mask file to unforce and mask split-usr for all
packages in generic images.

As a bonus, this also allows us to drop a customization we did in
sys-apps/policycoreutils package.