sys-kernel/coreos-modules: Add a workaround for BTF mismatches
There is probably a bug in the way we build our kernel modules or in
the toolchain that we are using for building them, but the generated
BTF information contains invalid data, which causes the BTF validator
to bail out. Without the MODULE_ALLOW_BTF_MISMATCH config enabled,
this results in a kernel module being not loaded, which results in
halted boot because of missing iscsi module.

The verification seems to be failing at line 3077 in kernel/bpf/btf.c:

https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/tree/kernel/bpf/btf.c?h=v6.1.21#n3077

which tells me that it checks if a name of a struct is a valid
identifier. I'd risk a guess that in our case the identifier is empty,
but I haven't debugged it that far yet.